Extract the Bootstrap method into a separate BootstrapRouting interface
The plan is to create a MultiIpfsRouter that can dispatch requests to multiple
ContentRouters, PeerRouters, and ValueStores. However, to do this, we need to be
able to call Bootstrap on these sub-routers.

By breaking this into a separate interface, these sub-routers can optionally
implement BootstrapRouting iff they need to be bootstrapped.