fix event channel close race
1. Instead of accepting a passed-in channel, create one (and return a read-only
channel).
2. Close the channel when the context is canceled.

Together, these allow us to safely close the event channel.

Technically, this is an API-breaking change but this is (a) an internal API
and (b) unavoidable (I think). That is, I'm pretty sure there's no way to fix
this without changing this method signature somehow (we need to somehow
synchronize closing of this channel with writing to it).

fixes https://github.com/ipfs/go-ipfs/issues/5616